renamed flex-market to ev-registry
this better confirms to the larger domain and project history
