+++ src/ev-registry/factories/nats-connection.ts
