+++ src/ev-registry/models/asset.ts
