+++ src/ev-registry/models/contracts/ev-registry.abi.ts
