+++ src/ev-registry/models/dids/did-factory.ts
