+++ src/ev-registry/models/dids/did.ts
