+++ src/index.ts
@@ -21,9 +21,9 @@ import { config } from "./config/config"
 import { locations } from "./data/locations"
 import { tokens } from "./data/tokens"
 import { Database } from "./database"
-import { FlexMarketClient } from "./flex-market/flex-market-client"
-import { EvRegistry } from "./flex-market/models/contracts/ev-registry"
-import { DIDFactory } from "./flex-market/models/dids/did-factory"
+import { FlexMarketClient } from "./ev-registry/flex-market-client"
+import { EvRegistry } from "./ev-registry/models/contracts/ev-registry"
+import { DIDFactory } from "./ev-registry/models/dids/did-factory"
 import { MockMonitorFactory } from "./models/mock-monitor-factory"
 import { IAssetIdentity, IDIDCache } from "./types"
